requirements.txt
pytest==6.2.2
pytest-cov==2.11.1
alembic==1.5.5
click==7.1.2
colorama==0.4.4
fastapi==0.73.0
h11==0.12.0
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.17.1
typing-extensions==4.0.0
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
sqlalchemy-pagination==0.0.2
requests==2.25.1
python-jose==3.2.0

